Improve bcrypt and jsonwebtoken utils
