+++ utils/bcrypt/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  encryptPassword: require('./encryptPassword'),
+  comparePassword: require('./comparePassword'),
+}
