+++ utils/jsonwebtoken/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  createToken: require('./createToken'),
+  verifyToken: require('./verifyToken'),
+}
